# openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Serial Commander API",
    "version": "1.0.0",
    "description": "API documentation for Serial Commander application"
  },
  "servers": [
    {
      "url": "http://localhost:2999",
      "description": "Development server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Thông tin tổng quan API",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "Thông tin trạng thái API",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RootSuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scenarios/myscenarios": {
      "get": {
        "summary": "Danh sách kịch bản của user",
        "tags": [
          "Scenario"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "legacy_array",
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            },
            "description": "Nếu \"1\", trả mảng JSON thuần (tương thích cũ). Mặc định envelope { message, scenarios }."
          }
        ],
        "responses": {
          "200": {
            "description": "Envelope chuẩn hoặc mảng (legacy)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ScenarioListSuccessResponse"
                    },
                    {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/scenarios/{scenarioId}": {
      "get": {
        "summary": "Chi tiết kịch bản (field kịch bản gộp root cùng message)",
        "tags": [
          "Scenario"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "scenarioId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Kịch bản",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioMergedResourceSuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scenarios/export/{scenarioId}": {
      "get": {
        "summary": "Xuất kịch bản (JSON, Content-Disposition attachment)",
        "tags": [
          "Scenario"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "scenarioId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioMergedResourceSuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/verify": {
      "post": {
        "summary": "Kiểm tra tính hợp lệ của payload kịch bản",
        "tags": [
          "Scenario"
        ],
        "responses": {
          "200": {
            "description": "Kết quả kiểm tra kịch bản",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioVerifySuccessResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/verify-file": {
      "post": {
        "summary": "Kiểm tra cú pháp file kịch bản JSON",
        "tags": [
          "Scenario"
        ],
        "responses": {
          "200": {
            "description": "Kết quả kiểm tra file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioVerifyFileSuccessResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/share/{shareCode}": {
      "get": {
        "summary": "Lấy kịch bản theo mã chia sẻ (public)",
        "tags": [
          "Scenario"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "shareCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Kịch bản (field gộp root)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioMergedResourceSuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/ops/sync-jobs": {
      "get": {
        "summary": "Tóm tắt SyncJobs (đồng bộ MySQL/Firestore) cho vận hành",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thống kê theo trạng thái + mẫu failed gần nhất",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncJobsOpsSummaryEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Không có hoặc token không hợp lệ"
          }
        }
      }
    },
    "/admin/ops/metrics": {
      "get": {
        "summary": "Counters process + gauges SyncJobs (JSON hoặc ?format=prometheus)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "format",
            "schema": {
              "type": "string",
              "enum": [
                "prometheus"
              ]
            },
            "description": "Nếu `prometheus` trả body text exposition"
          }
        ],
        "responses": {
          "200": {
            "description": "metrics JSON hoặc Prometheus text (`format=prometheus`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppOpsMetricsEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Đăng nhập người dùng",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đăng nhập thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sai tài khoản hoặc mật khẩu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Đăng ký tài khoản mới",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Đăng ký thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Email hoặc username đã tồn tại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/verify-email": {
      "post": {
        "summary": "Xác thực email bằng mã OTP",
        "tags": [
          "Authentication"
        ],
        "responses": {
          "200": {
            "description": "Xác thực thành công hoặc đã xác thực trước đó",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Mã không hợp lệ, hết hạn, hoặc input sai",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/resend-verification-code": {
      "post": {
        "summary": "Gửi lại mã xác thực email",
        "tags": [
          "Authentication"
        ],
        "responses": {
          "200": {
            "description": "Gửi mã thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Input không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/google/status": {
      "get": {
        "summary": "Trạng thái cấu hình Google OAuth",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/google": {
      "get": {
        "summary": "Đăng nhập bằng Google OAuth",
        "tags": [
          "Authentication"
        ],
        "responses": {
          "302": {
            "description": "Redirect đến Google OAuth"
          }
        }
      }
    },
    "/api/auth/google/callback": {
      "get": {
        "summary": "Google OAuth callback",
        "tags": [
          "Authentication"
        ],
        "responses": {
          "302": {
            "description": "Redirect về frontend với token"
          }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "summary": "Yêu cầu mã đặt lại mật khẩu",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email đã được gửi (nếu email tồn tại)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Input không hợp lệ hoặc tài khoản Google",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/verify-reset-code": {
      "post": {
        "summary": "Xác thực mã đặt lại mật khẩu",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "code"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mã hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResetCodeSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Mã không hợp lệ hoặc đã hết hạn",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "summary": "Đặt lại mật khẩu với mã xác thực",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "code",
                  "newPassword"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đặt lại mật khẩu thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Mã không hợp lệ hoặc mật khẩu không đủ dài",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message",
          "error"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "trace_id": {
            "type": "string",
            "description": "Mã bắc cầu request (trùng header X-Request-Id) khi có requestTraceMiddleware."
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Ví dụ RATE_LIMIT_EXCEEDED, NO_TOKEN, VALIDATION_FAILED."
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          }
        }
      },
      "ScenarioListSuccessResponse": {
        "type": "object",
        "required": [
          "message",
          "scenarios"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "scenarios": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "ScenarioMergedResourceSuccessResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "description": "Kịch bản: sendSuccess merge DTO — Name, Content, Banners, ... cùng cấp với message (client cũ vẫn đọc field kịch bản ở root).",
        "properties": {
          "message": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "LoginSuccessResponse": {
        "type": "object",
        "required": [
          "message",
          "token"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Đăng nhập thành công"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "RegisterSuccessResponse": {
        "type": "object",
        "required": [
          "message",
          "requireEmailVerification",
          "email",
          "emailSent"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "requireEmailVerification": {
            "type": "boolean"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "emailSent": {
            "type": "boolean"
          }
        }
      },
      "MessageSuccessResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "VerifyResetCodeSuccessResponse": {
        "type": "object",
        "required": [
          "message",
          "valid"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "valid": {
            "type": "boolean"
          }
        }
      },
      "ScenarioVerifySuccessResponse": {
        "type": "object",
        "required": [
          "message",
          "errors",
          "warnings"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ScenarioVerifyFileSuccessResponse": {
        "type": "object",
        "required": [
          "message",
          "valid",
          "errors",
          "warnings"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "valid": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "path": {
                  "type": "string"
                },
                "line": {
                  "type": "number"
                },
                "column": {
                  "type": "number"
                }
              }
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "path": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "RootSuccessResponse": {
        "type": "object",
        "required": [
          "message",
          "version",
          "endpoints",
          "status"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "trace_id": {
            "type": "string",
            "description": "Mã bắc cầu request (trùng header X-Request-Id)."
          },
          "version": {
            "type": "string"
          },
          "endpoints": {
            "type": "object"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "SyncJobsOpsSummary": {
        "type": "object",
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time"
          },
          "by_status": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "due_for_processing": {
            "type": "integer"
          },
          "failed_recent": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "operation_type": {
                  "type": "string"
                },
                "scenario_id": {
                  "type": "string"
                },
                "retry_count": {
                  "type": "integer"
                },
                "last_error": {
                  "type": "string",
                  "nullable": true
                },
                "modified_at": {
                  "type": "string",
                  "format": "date-time",
                  "nullable": true
                }
              }
            }
          }
        }
      },
      "SyncJobsOpsSummaryEnvelope": {
        "type": "object",
        "required": [
          "message",
          "summary"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/SyncJobsOpsSummary"
          }
        }
      },
      "AppOpsMetricsEnvelope": {
        "type": "object",
        "required": [
          "message",
          "metrics"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "metrics": {
            "type": "object",
            "properties": {
              "generated_at": {
                "type": "string",
                "format": "date-time"
              },
              "counters": {
                "type": "object",
                "additionalProperties": {
                  "type": "number"
                }
              },
              "gauges": {
                "type": "object",
                "additionalProperties": {
                  "type": "number"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}